+++ d60.Cirqus/Events/IEventStore.cs
@@ -16,7 +16,7 @@ namespace d60.Cirqus.Events
         /// <summary>
         /// Loads events for the specified aggregate root
         /// </summary>
-        IEnumerable<DomainEvent> Load(Guid aggregateRootId, long firstSeq = 0, long limit = int.MaxValue);
+        IEnumerable<DomainEvent> Load(Guid aggregateRootId, long firstSeq = 0);
 
         /// <summary>
         /// Streams all events with a global sequence number that is greater than or equal to the one given